TypeScript fixes
